--- a/microservices/download/package.json
+++ b/microservices/download/package.json
@@ -12,7 +12,6 @@
   "author": "confused-Techie",
   "license": "MIT",
   "dependencies": {
-    "express": "^4.18.1",
     "superagent": "^8.0.1"
   }
 }