This PR does just like the title says, and removes `Express` from the `download` microservice endpoint.

All initial handling of the web request is done within Node's Built in `HTTP` module.

This PR was very largely inspired by #49 as with this merged that would mean the `download` microservice is a dependency free solution.

While the PR linked does provide additional Speed and Size benefits, this PR comes with it's own as well.

Before:
  - Size: 4.03MB
  - Speed: 641ms - 515ms - 521ms

After:
  - Size: 2.13MB
  - Speed: 517ms - 532ms - 602ms
  
So while the performance in size means this microservice is nearly half the size as it was previously, unfortunately the speed advantage isn't that great. Although with that said, the microservice was still faster on average than when using Express.