@DeeDeeG 2023-01-19
I actually really like the code style of the new lines of `microservices/download/index.js`. I find it less confusing than the Express stuff. 👍

Can't say I fully understand all of it, but the code looks a lot more self-explanatory, easier to guess what is going on, and frankly less going on, which is _good_ IMO.