@confused-Techie 2023-01-19
@DeeDeeG To answer some of your questions here:

* Yes this does use HTTP which is not secure, but reason being is it doesn't need to be. SSL is terminated by our hosting environment before it hits the endpoint, so we don't have to worry about user facing SSL or encryption at all, it's all handled for us by the hosts reverse proxy. But we still do want to use `https` when reaching out from here as that's our own request and own responsibility, so you are good using `https` for your previous PR.
* As for the performance numbers, since I don't have access to the `time` utility on Windows and there is no good testing platform baked into this repo or microservice yet, I just testing by running the server locally and using Postman to hit the endpoint over and over, grabbing some average values. Not super scientific, but easy to test on your system if you'd like to give it a try.
* For the query parameters you are exactly right. They were an implementation from Express, with using `http` our query parameters are all within the URL itself, thats why you see me splitting up the actual path and query parameters in line 7 of `microservices/download/index.js`. So it is a bit more verbose to check our query parameters now. Which if we need to do much more of it, we could always create a dedicated function to handle this for us, so the code doesn't have to be repeated. But two uses side by side, seem simple enough to not warrant it just yet.
* But for `for...of` I will check that out. I mean really I can't imagine it's any large performance hit even if it is one. So simpler code is just a fine justification for using that instead. Since you are right, the current solution is a bit verbose.
* Lastly, you are totally correct, calling `end()` twice is pointless. The secondary `.end()` would likely never be reached, so it should be removed.

But thanks a ton for the review, and for thinking the code style is looking good over all. What can I say, after seeing your PR I was inspired to have a dependency-free microservice, as in the world of JavaScript and NPM that's so infrequent it's awesome. But yeah otherwise pretty familiar with this section, as I think it's only really been touched by me with @Spiker985 having written the GraphQL queries, so from my testing everything still works as we would expect. Although might be worth it to prior to bumping the version to add some testing. (Although suppose if it's done after the fact that doesn't really matter, since the testing wouldn't be in the final microservice anyway).

Again thanks for the review, I'll check out some of those simplification ideas you recommended, and see maybe if there's any more solid way to interact with query parameters to keep things less beefy there if possible. (Of course there's a million NPM modules to do this, but defeats my whole idea here. So ideally I can find some MIT licensed work within a good one and look at just a function implementation)