@confused-Techie 2023-01-21
Alright @DeeDeeG I've gone ahead and resolved the pointers you had (I know not needed but would rather get to it now)

And otherwise will go ahead and merge this one.

Thanks for your review, and approval! This microservice is now dependency free!